refactor(UserLevel): LevelModal 컴포넌트 분리
